refactor: Split and reworked DriveFileManager and DriveInfosManager in oder to make frozen more explicit
