+++ kDrive/UI/Controller/Favorite/FavoritesViewModel.swift
@@ -31,7 +31,7 @@ class FavoritesViewModel: FileListViewModel {
                                           selectedTabBarIcon: KDriveResourcesAsset.starFill,
                                           emptyViewType: .noFavorite,
                                           matomoViewPath: ["Favorite"])
-        let favoritesFakeRoot = driveFileManager.getManagedFile(from: DriveFileManager.favoriteRootFile)
+        let favoritesFakeRoot = driveFileManager.getLiveManagedFile(from: DriveFileManager.favoriteRootFile)
         super.init(configuration: configuration, driveFileManager: driveFileManager, currentDirectory: favoritesFakeRoot)
         files = AnyRealmCollection(AnyRealmCollection(favoritesFakeRoot.children.filter(NSPredicate(format: "isFavorite = true")))
             .filesSorted(by: sortType))
