+++ kDrive/UI/Controller/Menu/MySharesViewModel.swift
@@ -29,7 +29,7 @@ class MySharesViewModel: FileListViewModel {
                                           rootTitle: KDriveResourcesStrings.Localizable.mySharesTitle,
                                           emptyViewType: .noShared,
                                           matomoViewPath: [MatomoUtils.Views.menu.displayName, "MyShares"])
-        let mySharesFakeRoot = driveFileManager.getManagedFile(from: DriveFileManager.mySharedRootFile)
+        let mySharesFakeRoot = driveFileManager.getLiveManagedFile(from: DriveFileManager.mySharedRootFile)
         super.init(configuration: configuration, driveFileManager: driveFileManager, currentDirectory: mySharesFakeRoot)
         files = AnyRealmCollection(AnyRealmCollection(mySharesFakeRoot.children.filter(NSPredicate(format: "users.@count > 0")))
             .filesSorted(by: sortType))
