+++ kDriveCore/Data/Cache/DriveInfoManager/DriveInfosManager.swift
@@ -249,122 +249,4 @@ public final class DriveInfosManager: RealmAccessible {
     public static func getObjectId(driveId: Int, userId: Int) -> String {
         return "\(driveId)_\(userId)"
     }
-
-    public func getDrives(for userId: Int? = nil, sharedWithMe: Bool? = false, using realm: Realm? = nil) -> [Drive] {
-        let realm = realm ?? getRealm()
-        realm.refresh()
-        var realmDriveList = realm.objects(Drive.self)
-            .sorted(byKeyPath: "name", ascending: true)
-            .sorted(byKeyPath: "sharedWithMe", ascending: true)
-        if let userId {
-            let filterPredicate: NSPredicate
-            if let sharedWithMe {
-                filterPredicate = NSPredicate(format: "userId = %d AND sharedWithMe = %@", userId, NSNumber(value: sharedWithMe))
-            } else {
-                filterPredicate = NSPredicate(format: "userId = %d", userId)
-            }
-            realmDriveList = realmDriveList.filter(filterPredicate)
-        }
-        return Array(realmDriveList.map { $0.freeze() })
-    }
-
-    public func getFrozenDrive(id: Int, userId: Int) async -> Drive? {
-        var drive: Drive?
-        await realmTransaction.execute { realm in
-            drive = self.getFrozenDrive(id: id, userId: userId, using: realm)
-        }
-        return drive
-    }
-
-    public func getFrozenDrive(id: Int, userId: Int, using realm: Realm) -> Drive? {
-        let objectId = DriveInfosManager.getObjectId(driveId: id, userId: userId)
-        return getFrozenDrive(objectId: objectId, using: realm)
-    }
-
-    public func getFrozenDrive(objectId: String) async -> Drive? {
-        var drive: Drive?
-        await realmTransaction.execute { realm in
-            drive = self.getFrozenDrive(objectId: objectId, using: realm)
-        }
-        return drive
-    }
-
-    public func getFrozenDrive(objectId: String, using realm: Realm) -> Drive? {
-        getLiveDrive(objectId: objectId, using: realm)?.freeze()
-    }
-
-    public func getLiveDrive(objectId: String, using realm: Realm) -> Drive? {
-        guard let drive = realm.object(ofType: Drive.self, forPrimaryKey: objectId),
-              !drive.isInvalidated else {
-            return nil
-        }
-        return drive
-    }
-
-    public func getFrozenUsers(for driveId: Int, userId: Int) async -> [DriveUser] {
-        var driveUsers = [DriveUser]()
-        await realmTransaction.execute { realm in
-            driveUsers = self.getFrozenUsers(for: driveId, userId: userId, using: realm)
-        }
-        return driveUsers
-    }
-
-    public func getFrozenUsers(for driveId: Int, userId: Int, using realm: Realm) -> [DriveUser] {
-        guard let drive = getFrozenDrive(id: driveId, userId: userId, using: realm) else {
-            return []
-        }
-
-        let realmUserList = realm.objects(DriveUser.self).sorted(byKeyPath: "id", ascending: true)
-        let users = realmUserList.filter { drive.users.drive.contains($0.id) }
-        return users.map { $0.freeze() }
-    }
-
-    public func getFrozenUser(id: Int, using realm: Realm) async -> DriveUser? {
-        var driveUser: DriveUser?
-        await realmTransaction.execute { realm in
-            driveUser = self.getFrozenUser(id: id, using: realm)
-        }
-        return driveUser
-    }
-
-    public func getFrozenUser(id: Int, using realm: Realm) -> DriveUser? {
-        guard let user = realm.object(ofType: DriveUser.self, forPrimaryKey: id),
-              !user.isInvalidated else {
-            return nil
-        }
-        return user.freeze()
-    }
-
-    public func getFrozenTeams(for driveId: Int, userId: Int) async -> [Team] {
-        var teams = [Team]()
-        await realmTransaction.execute { realm in
-            teams = self.getFrozenTeams(for: driveId, userId: userId, using: realm)
-        }
-        return teams
-    }
-
-    public func getFrozenTeams(for driveId: Int, userId: Int, using realm: Realm) -> [Team] {
-        guard let drive = getFrozenDrive(id: driveId, userId: userId, using: realm) else {
-            return []
-        }
-
-        let realmTeamList = realm.objects(Team.self).sorted(byKeyPath: "id", ascending: true)
-        let filteredTeams = realmTeamList.filter { drive.teams.account.contains($0.id) }
-        return filteredTeams.map { $0.freeze() }
-    }
-
-    public func getFrozenTeam(id: Int) async -> Team? {
-        var team: Team?
-        await realmTransaction.execute { realm in
-            team = self.getFrozenTeam(id: id, using: realm)
-        }
-        return team
-    }
-
-    public func getFrozenTeam(id: Int, using realm: Realm) -> Team? {
-        guard let team = realm.object(ofType: Team.self, forPrimaryKey: id), !team.isInvalidated else {
-            return nil
-        }
-        return team.freeze()
-    }
 }
